calendar-agent/backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-socketio[asyncio]
pydantic
python-jose[cryptography]
httpx
python-dotenv
